build chatbox compnent
